[FIX] 게시판 수정
게시판 작성 수정
