+++ Java-Funiture-OnlineShop/src/main/java/com/example/funitureOnlineShop/home/HomeController.java
@@ -1,5 +1,6 @@
 package com.example.funitureOnlineShop.home;
 
+import com.example.funitureOnlineShop.Board.BoardDTO;
 import com.example.funitureOnlineShop.Board.BoardService;
 import com.example.funitureOnlineShop.category.CategoryResponse;
 import com.example.funitureOnlineShop.category.CategoryService;
@@ -10,6 +11,9 @@ import com.example.funitureOnlineShop.product.ProductService;
 import com.example.funitureOnlineShop.productComment.ProductCommentResponse;
 import com.example.funitureOnlineShop.productComment.ProductCommentService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -165,5 +169,25 @@ public class HomeController {
         model.addAttribute("sons", sons);
 
         return "menu";
+    }
+
+    @GetMapping("/boardPage")
+    public String paging(@PageableDefault(page = 1) Pageable pageable, Model model){
+        Page<BoardDTO> boards = boardService.paging(pageable);
+
+        int blockLimit = 3;
+        int startPage = (int)(Math.ceil((double)pageable.getPageNumber() / blockLimit) - 1) * blockLimit + 1;
+        int endPage = ((startPage + blockLimit - 1) < boards.getTotalPages()) ? (startPage + blockLimit - 1): boards.getTotalPages();
+
+        model.addAttribute("boardList", boards);
+        model.addAttribute("startPage", startPage);
+        model.addAttribute("endPage", endPage);
+
+        return "noticePage";
+    }
+
+    @GetMapping("/boardCreate")
+    public String boardCreateForm() {
+        return "createboard";
     }
 }
