+++ Java-Funiture-OnlineShop/src/main/resources/templates/index.html
@@ -225,7 +225,7 @@
 <section>
     <nav>
         <button class="btn" onclick="location.href='/productAll'">상품 페이지(준비중)</button>
-        <button class="btn" onclick="location.href='/board/'">게시판</button>
+        <button class="btn" onclick="location.href='/boardPage'">게시판</button>
     </nav>
     <span id="adminButton" style="display: none;">
         <button class="btn" onclick="location.href='/adminPage'">관리자 페이지</button>
